+++ poke-dex/package.json
@@ -14,6 +14,7 @@
     "react-dom": "^16.13.1",
     "react-js-pagination": "^3.0.3",
     "react-paginate": "^6.3.2",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.1"
   },
   "scripts": {
