+++ poke-dex/public/index.html
@@ -5,6 +5,7 @@
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
+    <script src="https://kit.fontawesome.com/5b96f0cf09.js" crossorigin="anonymous"></script>
     <meta
       name="description"
       content="Web site created using create-react-app"
