+++ poke-dex/src/scss/components/_navbar.scss
@@ -1,0 +1,47 @@
+@import '../settings/colors';
+
+.nav-bar {
+  width: 100%;
+  background-color: black;
+  opacity: 80%;
+  box-shadow: 0 0 5px black;
+
+}
+
+
+.nav-row{
+  height: 15vh;
+  margin-bottom: 1rem;
+  box-shadow: $headShadow;
+  @media (max-width: 560px) {
+    height: 15vh;
+  }
+
+  .header-container {
+    text-align: center;
+    color: white;
+    margin-top: 2rem;
+    p {
+      text-shadow: 0 0 12px $headShadow;
+    }
+
+    .nav-head {
+      font-size: 4rem;
+      font-weight: bold;
+      text-shadow: 0 0 12px $headShadow;
+    }
+    @media (max-width: 560px) {
+      .nav-head {
+        font-size: 2.5rem;
+      }
+    }
+  }
+
+  .header-container:hover {
+    p, .nav-head {
+      text-shadow: 0 0 18px $headShadow;
+    ;
+    }
+  }
+
+}
