+++ poke-dex/src/scss/settings/_colors.scss
@@ -1,0 +1,1 @@
+$headShadow: #f5eb82;
